address design comment
