+++ contract/src/nft/mod.rs
@@ -2,4 +2,3 @@ pub mod core;
 pub mod enumeration;
 pub mod internal;
 pub mod metadata;
-pub mod mint;
